Set the return value of length as size_t, not uint
Commit 1bab980 changed length member to return size_t,
but forget to edit the interface.

+++ src/ocean/util/container/pool/model/IPoolInfo.d
@@ -37,7 +37,7 @@ public interface IPoolInfo : IFreeListInfo, ILimitableInfo
 
      **************************************************************************/
 
-    uint length ( );
+    size_t length ( );
 
     /**************************************************************************
 
@@ -50,4 +50,3 @@ public interface IPoolInfo : IFreeListInfo, ILimitableInfo
 
     uint num_busy ( );
 }
-
